+++ src/shared_context.h
@@ -219,6 +219,6 @@ extern uint32_t eth2WithdrawalIndex;
 #endif
 
 void reset_app_context(void);
-const uint8_t *parseBip32(const uint8_t *, uint16_t *, bip32_path_t *);
+const uint8_t *parseBip32(const uint8_t *, uint8_t *, bip32_path_t *);
 
 #endif  // _SHARED_CONTEXT_H_
